Implement sorting and filtering options for activities
Extended the app with functionality to sort and filter activities based on different parameters such as distance, time, heart rate, and calories. Added methods to sort activities by these parameters and added new fields to the activity structure. Also implemented option to hide certain fields in the output.

+++ pkg/activity/activity.go
@@ -12,18 +12,18 @@ type Activity struct {
 	Time                   time.Duration
 	AvgHR                  int64
 	MaxHR                  int64
-	AvgRunCadence          float64
-	MaxRunCadence          float64
+	AvgRunCadence          int64
+	MaxRunCadence          int64
 	AvgPace                float64
 	BestPace               float64
-	TotalAscent            float64
-	TotalDescent           float64
+	TotalAscent            int64
+	TotalDescent           int64
 	AvgStrideLength        float64
 	AvgVerticalRatio       float64
 	AvgVerticalOscillation float64
 	AvgGroundContactTime   float64
 	TrainingStressScore    float64
-	AvgPower               float64
+	AvgPower               int64
 	Grit                   float64
 	Flow                   float64
 	TotalStrokes           float64
@@ -35,7 +35,7 @@ type Activity struct {
 	SurfaceInterval        float64
 	Decompression          float64
 	BestLapTime            time.Time
-	NumberofLaps           float64
+	NumberofLaps           int64
 	MaxTemp                float64
 	AvgResp                float64
 	MinResp                float64
@@ -47,6 +47,6 @@ type Activity struct {
 	MaxStress              float64
 	MovingTime             float64
 	ElapsedTime            float64
-	MinElevation           float64
-	MaxElevation           float64
+	MinElevation           int64
+	MaxElevation           int64
 }
